pyproject.toml
[project]
name = "vb6-antlr"
version = "0.1.0"
description = "VB6 parser to JSON converter"
authors = [
    {name = "Malvoison",email = "allawyer66@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "antlr4-python3-runtime (>=4.13.2,<5.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)"
]


[tool.poetry]
packages = [{ include = "vb6_antlr", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = [
    "src/vb6_grammar/grammars",
]

[tool.mypy]
python_version = "3.11"
[[tool.mypy.overrides]]
module = ["vb6_grammar.grammars", "vb6_grammar.grammars.*"]
ignore_errors = true
